generic_analyst_agent/requirements.txt
langchain==0.2.16
langchain-core==0.2.39
langgraph==0.2.27
pandas==2.2.2
python-dotenv==1.0.1
langchain-groq==0.1.5
google-api-python-client==2.143.0
google-generativeai==0.8.3
pydantic==2.8.2
pyarrow==16.1.0
fastapi==0.110.3
uvicorn==0.30.6
streamlit==1.37.1
